requirements.txt
httplib2==0.9.2
PyYAML==3.12
requests==2.12.4
